+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Unit tests
 - Doxygen documentation (incomplete)
 - "Skip" added to Alarm "Enabled" options
+- Timeout for alarms - the alarm stops if it is active for too long
 
 ### Changed
 - Better LCD backlight control